--- rust-connectors/sinks/postgres/tests/data.sql
@@ -1,76 +1,0 @@
-CREATE TABLE big_table (
-	col_1 bigint,
-	col_2 bigserial,
-	col_3 bit,
-	col_4 boolean,
-	col_5 box,
-	col_6 bytea,
-	col_7 character,
-	col_8 cidr,
-	col_9 circle,
-	col_10 date,
-	col_11 float8,
-	col_12 inet,
-	col_13 integer,
-	col_14 interval,
-	col_15 json,
-	col_16 jsonb,
-	col_17 line,
-	col_18 lseg,
-	col_19 macaddr,
-	col_20 money,
-	col_21 numeric,
-	col_22 path,
-	col_23 pg_lsn,
-	col_24 point,
-	col_25 polygon,
-	col_26 real,
-	col_27 smallint,
-	col_28 smallserial,
-	col_29 text,
-	col_30 time,
-	col_31 timestamp,
-	col_32 tsquery,
-	col_33 tsvector,
-	col_34 txid_snapshot,
-	col_35 uuid,
-	col_36 xml
-);
-
-CREATE TABLE better_big_table (
-    bigint_col bigint,
-    bigserial_col bigserial,
-    bit_col bit,
-    boolean_col boolean,
-    box_col box,
-    bytea_col bytea,
-    character_col character,
-    cidr_col cidr,
-    circle_col circle,
-    date_col date,
-    inet_col inet,
-    integer_col integer,
-    interval_col interval,
-    json_col json,
-    jsonb_col jsonb,
-    line_col line,
-    lseg_col lseg,
-    macaddr_col macaddr,
-    money_col money,
-    numeric_col numeric,
-    path_col path,
-    pg_lsn_col pg_lsn,
-    point_col point,
-    polygon_col polygon,
-    real_col real,
-    smallint_col smallint,
-    smallserial_col smallserial,
-    text_col text,
-    time_col time,
-    timestamp_col timestamp,
-    tsquery_col tsquery,
-    tsvector_col tsvector,
-    txid_snapshot_col txid_snapshot,
-    uuid_col uuid,
-    xml_col xml
-);
